Ad nodecli provider
