+++ wallet/provider.go
@@ -1,0 +1,19 @@
+package wallet
+
+type Utxo struct {
+	Address Address
+	TxId    TxId
+	Amount  uint64
+	Index   uint64
+}
+
+type NodeTip struct {
+	Epoch uint64
+	Block uint64
+	Slot  uint64
+}
+
+type Provider interface {
+	QueryUtxos(Address) ([]Utxo, error)
+	QueryTip() (NodeTip, error)
+}
